+++ SVRichPresence/RichPresenceMod.cs
@@ -1,4 +1,6 @@
 ﻿using DiscordRPC;
+using DiscordRPC.Message;
+using Microsoft.Xna.Framework;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
@@ -6,7 +8,6 @@ using StardewValley;
 using StardewValley.Menus;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using Constants = StardewModdingAPI.Constants;
 using LogLevel = StardewModdingAPI.LogLevel;
@@ -14,11 +15,15 @@ using Utility = StardewValley.Utility;
 
 namespace SVRichPresence {
 	public class RichPresenceMod : Mod {
-		private const string clientId = "444517509148966923";
-		private const string steamId = "413150";
+		private static readonly Color blurple = new Color(114, 137, 218);
+		private static readonly string clientId = "444517509148966923";
+		private static readonly string steamId = "413150";
+		private readonly Random rand = new Random();
 		private ModConfig config = new ModConfig();
 		private IRichPresenceAPI api;
 		private DiscordRpcClient client;
+		private readonly JoinRequestMessage[] requests = new JoinRequestMessage[ushort.MaxValue + 1];
+		private ushort lastRequestID = 0;
 
 		public override void Entry(IModHelper helper) {
 #if DEBUG
@@ -52,26 +57,69 @@ namespace SVRichPresence {
 
 			api = new RichPresenceAPI(this);
 			client = new DiscordRpcClient(clientId,
+				autoEvents: false,
 				logger: new MonitorLogger(Monitor));
 			client.RegisterUriScheme(steamId);
 			client.OnReady += (sender, e) => {
-				Monitor.Log("Connected to Discord: " + GetDiscordTag(e.User, true), LogLevel.Info);
+				Monitor.Log("Connected to Discord: " + e.User.ToString(), LogLevel.Info);
 			};
 			client.OnJoin += (sender, args) => {
-				Monitor.Log("Attempting to join game", LogLevel.Info);
+				Monitor.Log("Attempting to join game: " + args.Secret, LogLevel.Info);
 				JoinGame(args.Secret);
 			};
-			client.OnJoinRequested += (sender, args) => {
-				string tag = GetDiscordTag(args.User);
-				Monitor.Log(tag + " is requesting to join your game.", LogLevel.Alert);
-				Monitor.Log("You can respond to this request from Discord or its overlay.", LogLevel.Info);
-				Game1.chatBox.addInfoMessage(tag + " is requesting to join your game. You can respond to this request in Discord Overlay.");
+			client.OnJoinRequested += (sender, msg) => {
+				string name = msg.User.Username;
+				string tag = msg.User.ToString();
+				ushort id = (ushort)rand.Next(ushort.MinValue, ushort.MaxValue);
+				requests[id] = msg;
+				lastRequestID = id;
+				string hex = id.ToString("X");
+				Monitor.Log(tag + " wants to join your game via Discord.", LogLevel.Alert);
+				Monitor.Log("To respond type \"discord " + hex + " yes/no\" or just \"discord yes/no\"", LogLevel.Info);
+				Game1.chatBox.addMessage(name + " wants to join your game via Discord.\nTo respond check the console or use Discord or its overlay.", blurple);
 			};
 			client.Initialize();
 			client.SetSubscription(EventType.Join | EventType.JoinRequest);
 
-			Helper.ConsoleCommands.Add("DiscordRP_TestJoin",
-				"Command for debugging.",
+			#region Console Commands
+			Helper.ConsoleCommands.Add("discord",
+				"Respond to a Discord join request.",
+				(command, args) => {
+					// Yes, I know this code is a mess.
+					switch (args[0].ToLower()) {
+						case "yes":
+						case "y":
+							Respond(lastRequestID, true);
+							break;
+						case "no":
+						case "n":
+							Respond(lastRequestID, false);
+							break;
+						default:
+							try {
+								var id = ushort.Parse(args[0], System.Globalization.NumberStyles.HexNumber);
+								switch (args[1].ToLower()) {
+									case "yes":
+									case "y":
+										Respond(id, true);
+										break;
+									case "no":
+									case "n":
+										Respond(id, false);
+										break;
+									default:
+										Monitor.Log("Invalid response.", LogLevel.Error);
+										break;
+								}
+							} catch (Exception) {
+								Monitor.Log("Invalid request ID.", LogLevel.Error);
+							}
+							break;
+					}
+				}
+			);
+			Helper.ConsoleCommands.Add("DiscordRP_Join",
+				"Join a co-op game via invite code.",
 				(string command, string[] args) => {
 					JoinGame(string.Join(" ", args));
 				}
@@ -136,6 +184,7 @@ namespace SVRichPresence {
 					Monitor.Log(string.Join(Environment.NewLine, output), LogLevel.Info);
 				}
 			);
+			#endregion
 			LoadConfig();
 
 			Helper.Events.Input.ButtonReleased += HandleButton;
@@ -152,6 +201,8 @@ namespace SVRichPresence {
 			};
 
 			ITagRegister tagReg = api.GetTagRegister(this);
+
+			#region Default Tags
 
 			tagReg.SetTag("Activity", () => api.GamePresence);
 			tagReg.SetTag("ModCount", () => Helper.ModRegistry.GetAll().Count());
@@ -205,12 +256,22 @@ namespace SVRichPresence {
 				Context.IsMultiplayer && Context.IsMainPlayer ? "Hosting" : "Playing", true);
 			tagReg.SetTag("GameNoun", () => Context.IsMultiplayer ? "Co-op" : "Solo", true);
 			tagReg.SetTag("GameInfo", () => api.GetTag("GameVerb") + " " + api.GetTag("GameNoun"), true);
+			#endregion
+		}
+
+		private void Respond(ushort id, Boolean response) {
+			var request = requests[id];
+			if (request == null)
+				Monitor.Log("Request ID doesn't exist.", LogLevel.Error);
+			client.Respond(request, response);
+			Monitor.Log("Responding to join request. This may fail if 30 seconds have passed.", LogLevel.Info);
+			requests[id] = null;
 		}
 
 		private void JoinGame(string inviteCode) {
-			object lobby = Program.sdk.Networking.GetLobbyFromInviteCode(inviteCode);
-			if (lobby == null) return;
 			Game1.ExitToTitle(() => {
+				object lobby = Program.sdk.Networking.GetLobbyFromInviteCode(inviteCode);
+				if (lobby == null) return;
 				TitleMenu.subMenu = new FarmhandMenu(Program.sdk.Networking.CreateClient(lobby));
 			});
 		}
@@ -230,6 +291,7 @@ namespace SVRichPresence {
 		}
 
 		private void LoadConfig() => config = Helper.ReadConfig<ModConfig>();
+
 		private void SaveConfig() => Helper.WriteConfig(config);
 
 		private Timestamps timestampSession;
@@ -238,49 +300,52 @@ namespace SVRichPresence {
 		private void SetTimestamp() => timestampFarm = Timestamps.Now;
 
 		private void DoUpdate(object sender, UpdateTickedEventArgs e) {
-			try {
-				if (e.IsMultipleOf(30))
-					client.SetPresence(GetActivity());
-			} catch { }
+			client.Invoke();
+			if (e.IsMultipleOf(30))
+				client.SetPresence(GetPresence());
 		}
 
 		private MenuPresence Conf => !Context.IsWorldReady ?
 			config.MenuPresence : config.GamePresence;
 
-		private RichPresence GetActivity() {
-			var presence = new RichPresence() {
+		private RichPresence GetPresence() {
+			var presence = new RichPresence {
 				Details = api.FormatText(Conf.Details),
-				State = api.FormatText(Conf.State),
-				Assets = {
-					LargeImageKey = "default_large",
-					LargeImageText = api.FormatText(Conf.LargeImageText),
-					SmallImageText = api.FormatText(Conf.SmallImageText)
-				}
+				State = api.FormatText(Conf.State)
 			};
-			if (Conf.ForceSmallImage || presence.Assets.SmallImageText.Length > 0)
-				presence.Assets.SmallImageKey = "default_small";
+			var assets = new Assets {
+				LargeImageKey = "default_large",
+				LargeImageText = api.FormatText(Conf.LargeImageText),
+				SmallImageText = api.FormatText(Conf.SmallImageText)
+			};
+			if (Conf.ForceSmallImage || assets.SmallImageText?.Length > 0)
+				assets.SmallImageKey = "default_small";
 
 			if (Context.IsWorldReady) {
 				var conf = (GamePresence)Conf;
 				if (conf.ShowSeason)
-					presence.Assets.LargeImageKey = $"{Game1.currentSeason}_{FarmTypeKey()}";
+					assets.LargeImageKey = $"{Game1.currentSeason}_{FarmTypeKey()}";
 				if (conf.ShowWeather)
-					presence.Assets.SmallImageKey = "weather_" + WeatherKey();
+					assets.SmallImageKey = "weather_" + WeatherKey();
 				if (conf.ShowPlayTime)
 					presence.Timestamps = timestampFarm;
 				if (Context.IsMultiplayer && conf.AllowAskToJoin)
 					try {
-						presence.Party.ID = Game1.MasterPlayer.UniqueMultiplayerID.ToString();
-						presence.Party.Size = Game1.numberOfPlayers();
-						presence.Party.Max = Game1.getFarm().getNumberBuildingsConstructed("Cabin") + 1;
-						presence.Secrets.JoinSecret = Game1.server.getInviteCode();
+						presence.Party = new Party {
+							ID = Game1.MasterPlayer.UniqueMultiplayerID.ToString(),
+							Size = Game1.numberOfPlayers(),
+							Max = Game1.getFarm().getNumberBuildingsConstructed("Cabin") + 1
+						};
+						presence.Secrets = new Secrets {
+							JoinSecret = Game1.server.getInviteCode()
+						};
 					} catch { }
 			}
 
 			if (config.ShowGlobalPlayTime)
 				presence.Timestamps = timestampSession;
 
-			return presence;
+			return presence.WithAssets(assets);
 		}
 
 		private string FarmTypeKey() {
@@ -314,13 +379,6 @@ namespace SVRichPresence {
 			if (Game1.isFestival())
 				return "festival";
 			return "sunny";
-		}
-
-		private string GetDiscordTag(User user, bool includeId = false) {
-			string ret = user.ToString();
-			if (includeId)
-				ret += " (" + user.ID + ")";
-			return ret;
 		}
 	}
 }