+++ SVRichPresence/CHANGELOG.md
@@ -5,11 +5,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.3.1] - 2019-12-05
+### Added
+- `discord` console command to respond to Ask to Join requests via console.
+
+### Changed
+- `DiscordRP_TestJoin` changed to `DiscordRP_Join`
+- Discord RPC only logs now in verbose mode.
+
 ### Removed
-- Once again removed Mac and Linux auto-setup.
+- Support for Mac and Linux.
 
 ### Fixed
 - Fixed crashing from Discord Game SDK by changing libraries.
+- Fixed joining games.
 
 ## [2.3.0] - 2019-12-05
 ### Added
@@ -111,7 +121,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Removed Discord event handlers due to them not triggering.
 
-[Unreleased]: https://github.com/FayneAldan/SVRichPresence/compare/v2.3.0...HEAD
+[Unreleased]: https://github.com/FayneAldan/SVRichPresence/compare/v2.3.1...HEAD
+[2.3.1]: https://github.com/FayneAldan/SVRichPresence/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/FayneAldan/SVRichPresence/compare/2.3.0-beta1...v2.3.0
 [2.3.0-beta1]: https://github.com/FayneAldan/SVRichPresence/compare/2.2.2...2.3.0-beta1
 [2.2.2]: https://github.com/FayneAldan/SVRichPresence/compare/2.2.1...2.2.2